Most modules complete, only database entry scripts to go

+++ main.csh
@@ -8,7 +8,7 @@
 ### runs sub csh routines to command individual processes (eg, reduce_image,
 ### radial_velocity, spectype etc)
 
-### usage: ./main.csh
+### usage: ./main.csh [file_path file_name]
 
 #############
 ### To Do ###
@@ -18,11 +18,17 @@
 ### Load config_file ###
 ########################
 
-### Define file_path
-set file_path = `grep FILE_PATH config_file | awk '{print $2}'`
 
-### Define file_name
-set file_name = `grep FILE_NAME config_file | awk '{print $2}'`
+### Define file_path and file_name
+if $1 == "" then
+    set file_path = `grep FILE_PATH config_file | awk '{print $2}'`
+    set file_name = `grep FILE_NAME config_file | awk '{print $2}'`
+else
+    set file_path = $1
+    set file_name = $2
+endif
+
+echo Beginning reduction of $file_name
 
 ####################
 ### Setup folder ###
@@ -30,9 +36,12 @@ set file_name = `grep FILE_NAME config_file | awk '{print $2}'`
 
 ### Remove previously reduced files - if DELETE_ALL is turned on
 set delete_all = `grep DELETE_ALL config_file | awk '{print $2}'`
-if ($delete_all == true) then 
+if ($delete_all == true) then
+    if ($1 == "") then
+    echo Deleting all previous reductions
     rm -rf $file_path/temp/*
     rm -rf $file_path/reduced/*
+    endif
 endif
 
 ### Remove previous reductions of FILE_NAME
@@ -64,3 +73,5 @@ endif
 ### Remove the iraf log file it stores in this folder.
 rm logfile
 rm tmp*.fits
+
+rm -f $file_path/temp/*
